requirements.txt
# Core
streamlit>=1.24
langchain>=1.0.8
langchain-groq>=1.0.1
groq>=0.30.0,<1.0.0
requests>=2.32.5,<3.0
pandas
Pillow
python-dotenv
